Move resource detectors test to observabilityManager.test.ts
